+++ gems/aws-sdk-core/lib/aws-sdk-core/xml/error_handler.rb
@@ -26,8 +26,7 @@ module Aws
         end
         context[:request_id] = request_id(body)
         errors_module = context.client.class.errors_module
-        error_class = errors_module.error_class(code).new(context, message, data)
-        error_class
+        errors_module.error_class(code).new(context, message, data)
       end
 
       def extract_error(body, context)
@@ -48,14 +47,32 @@ module Aws
             # match modeled shape name
             error_shape_code = rule.shape['error']['code'] if rule.shape['error']
             match = (code == error_shape_code || code == rule.shape.name)
-            if match && rule.shape.members.any?
-              data = Parser.new(rule).parse(context.http_response.body_contents)
-            end
+            next unless match && rule.shape.members.any?
+
+            data = parse_error_data(rule, context.http_response.body_contents)
+            # supporting HTTP bindings
+            apply_error_headers(rule, context, data)
           end
         end
         data
       rescue Xml::Parser::ParsingError
         EmptyStructure.new
+      end
+
+      def parse_error_data(rule, body)
+        # errors may nested under <Errors><Error>structure_data</Error></Errors>
+        # Or may be flat and under <Error>structure_data</Error>
+        body = body.tr("\n", '')
+        if matches = body.match(/<Error>(.+?)<\/Error>/)
+          Parser.new(rule).parse("<#{rule.shape.name}>#{matches[1]}</#{rule.shape.name}>")
+        else
+          EmptyStructure.new
+        end
+      end
+
+      def apply_error_headers(rule, context, data)
+        headers = Aws::Rest::Response::Headers.new(rule)
+        headers.apply(context.http_response, data)
       end
 
       def error_code(body, context)