+++ gems/aws-sdk-core/spec/aws/xml/builder_spec.rb
@@ -221,16 +221,18 @@ module Aws
         XML
       end
 
-      it 'supports flat list with locationName traits' do
+      it 'does not support member locationName trait in flattened lists' do
         shapes['IntegerList']['flattened'] = true
+        # locationName trait on targeted member is ignored when serializing
+        # serializing flattened lists in structures
         shapes['IntegerList']['member']['locationName'] = 'Number'
-        expect(xml(string:'abc', number_list:[1,2,3])).to eq(<<-XML)
-<xml>
-  <Number>1</Number>
-  <Number>2</Number>
-  <Number>3</Number>
-  <String>abc</String>
-</xml>
+        expect(xml(string: 'abc', number_list: [1, 2, 3])).to eq(<<~XML)
+          <xml>
+            <NumberList>1</NumberList>
+            <NumberList>2</NumberList>
+            <NumberList>3</NumberList>
+            <String>abc</String>
+          </xml>
         XML
       end
 
@@ -281,11 +283,13 @@ module Aws
           XML
         end
 
-        it 'applies xml namespaces to any shape' do
-          ns = { 'xmlNamespace' => {
-            'prefix' => 'xsi',
-            'uri' => 'http://xmlns.com/uri'
-          }}
+        it 'applies xml namespaces only to top level shapes' do
+          ns = {
+            'xmlNamespace' => {
+              'prefix' => 'xsi',
+              'uri' => 'http://xmlns.com/uri'
+            }
+          }
           shapes['StringShape'].update(ns)
           shapes['StructureShape']['members']['Nested'].update(ns)
           params = {
@@ -293,12 +297,12 @@ module Aws
               string: 'abc'
             }
           }
-          expect(xml(params)).to eq(<<-XML)
-<xml>
-  <Nested xmlns:xsi="http://xmlns.com/uri">
-    <String xmlns:xsi="http://xmlns.com/uri">abc</String>
-  </Nested>
-</xml>
+          expect(xml(params)).to eq(<<~XML)
+            <xml>
+              <Nested xmlns:xsi="http://xmlns.com/uri">
+                <String>abc</String>
+              </Nested>
+            </xml>
           XML
         end
       end